[module] Verplaats forum naar een module
