+++ htdocs/.htaccess
@@ -34,7 +34,7 @@ RewriteRule ^profiel			index.php?c=Profiel [QSA,L]
 RewriteRule ^groepen			index.php?c=GroepenRouter [L]
 RewriteRule ^maaltijden			index.php?c=MaalcieRouter [L]
 RewriteRule ^corvee				index.php?c=MaalcieRouter [L]
-RewriteRule ^forum				index.php?c=Forum [L]
+RewriteRule ^forum				index.php?m=forum&c=Forum [L]
 RewriteRule ^agenda				index.php?c=Agenda [L]
 RewriteRule ^courant			index.php?m=courant&c=Courant [L]
 RewriteRule ^fotoalbum			index.php?c=FotoAlbum [L]
