+++ lib/forum/model/ForumDradenModel.class.php
@@ -2,6 +2,8 @@
 use CsrDelft\Orm\CachedPersistenceModel;
 use CsrDelft\Orm\Persistence\Database;
 
+require_once 'forum/model/entity/ForumDraad.class.php';
+
 /**
  * ForumDradenModel.class.php
  *
