+++ lib/forum/model/ForumModel.class.php
@@ -5,13 +5,15 @@ use CsrDelft\Orm\Persistence\Database;
 
 require_once 'model/Paging.interface.php';
 
-require_once 'model/forum/ForumDelenModel.class.php';
-require_once 'model/forum/ForumDradenReagerenModel.class.php';
-require_once 'model/forum/ForumDradenGelezenModel.class.php';
-require_once 'model/forum/ForumDradenVerbergenModel.class.php';
-require_once 'model/forum/ForumDradenVolgenModel.class.php';
-require_once 'model/forum/ForumDradenModel.class.php';
-require_once 'model/forum/ForumPostsModel.class.php';
+require_once 'forum/model/ForumDelenModel.class.php';
+require_once 'forum/model/ForumDradenReagerenModel.class.php';
+require_once 'forum/model/ForumDradenGelezenModel.class.php';
+require_once 'forum/model/ForumDradenVerbergenModel.class.php';
+require_once 'forum/model/ForumDradenVolgenModel.class.php';
+require_once 'forum/model/ForumDradenModel.class.php';
+require_once 'forum/model/ForumPostsModel.class.php';
+
+require_once 'forum/model/entity/ForumCategorie.class.php';
 
 /**
  * ForumModel.class.php
