+++ lib/forum/model/entity/ForumCategorie.class.php
