+++ lib/forum/view/ForumView.class.php
